Move scrapy/contrib/loader to scrapy/loader
