+++ scrapy/loader/__init__.py
