+++ scrapy/loader/common.py
