+++ scrapy/loader/processor.py
