+++ tests/py3-ignores.txt
@@ -6,7 +6,7 @@ tests/test_commands.py
 tests/test_command_version.py
 tests/test_exporters.py
 tests/test_linkextractors.py
-tests/test_contrib_loader.py
+tests/test_loader.py
 tests/test_crawl.py
 tests/test_crawler.py
 tests/test_downloader_handlers.py
